Move disable events to factory
